refract code into submodules; update demo&layer card models
