+++ src/app/demos/point-cloud/point-cloud-view.component.ts
@@ -1,0 +1,19 @@
+import { Component } from '@angular/core';
+import { PointCloudDemoService } from './point-cloud-demo.service';
+import { POINTCLOUDDEMO } from './point-cloud-demo';
+import * as chroma from 'chroma-js';
+
+@Component({
+  selector: 'gd-test',
+  templateUrl: './point-cloud-view.component.html'
+})
+
+export class PointCloudViewComponent {
+  layer: L.Layer;
+  mapOptions = {
+    zoom: POINTCLOUDDEMO.zoom,
+    center: POINTCLOUDDEMO.center,
+    layers: POINTCLOUDDEMO.baseLayer,
+  };
+  layerCards = POINTCLOUDDEMO.layers;
+}
