+++ src/app/demos/potsdam/potsdam-view.component.html
@@ -1,0 +1,1 @@
+<gd-map-view [cards]='layerCards' [mapOptions]='mapOptions'></gd-map-view>
