+++ src/app/demos/potsdam/potsdam.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { CommonModule } from '@angular/common';
+import { PotsdamViewComponent } from './potsdam-view.component';
+
+import { SharedModule } from '../../shared/shared.module';
+import { PotsdamDemoService } from './potsdam-demo.service';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    SharedModule,
+    RouterModule.forChild([
+      { path: '', component: PotsdamViewComponent },
+    ])
+  ],
+  declarations: [PotsdamViewComponent],
+  providers: [PotsdamDemoService]
+})
+
+export class PotsdamModule { }
