+++ src/app/pages/dashboard/dashboard.component.html
@@ -3,9 +3,9 @@
     Demos and Case Studies
   </header>
   <nav class="card-tiles">
-    <a *ngFor="let card of cards" class="tile-item" [routerLink]="['/mapview', card.model]" routerLinkActive="active">
-      <div class="title">{{card.title}}</div>
-      <img class="thumb" [src]="card.thumb" alt="Loading...">
+    <a *ngFor="let demo of demos" class="tile-item" [routerLink]="['/', demo.name]" routerLinkActive="active">
+      <div class="title">{{demo.title}}</div>
+      <img class="thumb" [src]="demo.thumb" alt="Loading...">
     </a>
   </nav>
 </div>
