+++ src/app/shared/components/map-view/sidebar-section/layer-card/layer-card.component.html
