+++ src/app/shared/components/map-view/sidebar-section/layer-card/param-item/param-item.component.ts
