+++ src/app/shared/components/map-view/sidebar-section/sidebar-header/sidebar-header.component.html
