+++ src/app/shared/models/layer-card.d.ts
@@ -1,7 +1,4 @@
 export interface LayerCard {
-  model: string;
-  title: string;
-  thumb: string;
   info: {
     name: string;
     title: string;
@@ -15,8 +12,6 @@ export interface LayerCard {
     prtext?: string[];
     range?: number;
     palettes?: string[];
-    zoom: number;
-    center: number[];
   };
   values?: number[] | string[];
   show: boolean;
